src/pyproject.toml
[project]
name = "messenger"
version = "0.0.1"
description = ""
authors = [
    {name = "MuratovER cooperative.entr@gmail.com"}
]
readme = "../README.md"
requires-python = "==3.13.2"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.11.1,<3.0.0)",
    "sqlalchemy[mypy] (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "inflection (>=0.5.1,<0.6.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "httptools (>=0.6.4,<0.7.0)",
    "uvloop (>=0.21.0,<0.22.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.3,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "types-requests (>=2.32.0.20250328,<3.0.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "fastapi-pagination (>=0.12.34,<0.13.0)",
    "orjson (>=3.10.16,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "psycopg (>=3.2.6,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.11.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-factoryboy = "^2.7.0"
types-requests = "^2.32.0.20250328"
types-passlib = "^1.7.7.20250401"

